% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings_table.R
\name{genre_distribution_plot}
\alias{genre_distribution_plot}
\title{Generating a parameterized genre distribution plot.}
\usage{
genre_distribution_plot(tv_data, selected_genre)
}
\arguments{
\item{tv_data}{The TV ratings data.}

\item{selected_genre}{The number of top genres to include in the plot (constant number).}
}
\value{
The plot for the genre distribution.
}
\description{
This function generates the genre distribution plot based on the selected
top genres.
}
