% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genre_plot.R
\name{genre_plot}
\alias{genre_plot}
\title{Generating a parameterized genre distribution plot.}
\usage{
genre_plot(tv_data, selected_genre)
}
\arguments{
\item{tv_data}{The TV ratings data.}

\item{selected_genre}{The number of top genres to include in the plot (constant number).}
}
\value{
The plot for the genre distribution.
}
\description{
This function generates the genre distribution plot based on the selected
top genres.
}
\examples{
# example code

genre_plot(tv_ratings, selected_genre = 5)

tv_ratings is the dataset while selected_genre = 5 shows the top 5 genres in the plot.

}
