% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings_summary.R
\name{summary_tvrating}
\alias{summary_tvrating}
\title{Calculating the Five-Number Summary for TV Ratings}
\usage{
summary_tvrating(tvrating)
}
\arguments{
\item{tv_data}{is the data frame that contains the TV ratings data.}
}
\value{
is the data frame with two columns: "Statistic" (the name of the statistic) and "Value".
}
\description{
This calculates the five-number summary (minimum, Q1, median, Q3, maximum) for the "av_rating" column in the TV ratings dataset.
}
\details{
If the input is a numerical vector of at least 5 values, it returns a data frame with two columns in which the first column consists the names of the values (minimum, Q1, median, Q3, maximum) while the second column returns the five number summary values.

If the input is a numerical vector of less than 5 values, it returns NULL.
}
