% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{tv_ratings}
\alias{tv_ratings}
\title{TV Ratings Data Set}
\format{
A data frame with 2,266 rows and 7 columns:
\describe{
\item{titleId}{Unique identifier for TV show titles. Character.}
\item{seasonNumber}{Season number of the TV show. Numeric.}
\item{title}{Title of the TV show. Character.}
\item{date}{Date of the TV show. Date.}
\item{av_rating}{Average viewer rating of the TV show. Numeric.}
\item{share}{Viewer share of the TV show. Numeric.}
\item{genres}{Genres of the TV show. Character.}
}
}
\description{
This data set contains information about TV ratings.
}
\keyword{datasets}
